fix: table of contents bug
