+++ packages/ui/app/src/atoms/index.ts
@@ -17,6 +17,7 @@ export * from "./sidebar";
 export * from "./store";
 export * from "./stream";
 export * from "./theme";
+export * from "./toc";
 export * from "./types";
 export * from "./utils";
 export * from "./viewport";
