+++ packages/ui/app/src/components/TableOfContents.tsx
@@ -1,7 +1,8 @@
 import cn from "clsx";
 import { useAtomValue } from "jotai";
-import { CSSProperties, useEffect, useMemo, useState } from "react";
-import { ANCHOR_ATOM } from "../atoms";
+import { CSSProperties } from "react";
+import { useMemoOne } from "use-memo-one";
+import { createAnchorInViewAtom, createTocAtom, type TableOfContentsItem } from "../atoms";
 import { FernLink } from "./FernLink";
 
 export declare namespace TableOfContents {
@@ -17,9 +18,6 @@ export declare namespace TableOfContents {
         tableOfContents: TableOfContentsItem[];
     }
 }
-
-let anchorJustSet = false;
-let anchorJustSetTimeout: number;
 
 function TableOfContentsItem({
     text,
@@ -98,70 +96,9 @@ function TableOfContentsList({
 }
 
 export const TableOfContents: React.FC<TableOfContents.Props> = ({ className, tableOfContents, style }) => {
-    const currentPathAnchor = useAtomValue(ANCHOR_ATOM);
-
-    const allAnchors = useMemo(() => getAllAnchorStrings(tableOfContents), [tableOfContents]);
-
-    const [anchorInView, setAnchorInView] = useState(currentPathAnchor ?? allAnchors[0]);
-
-    useEffect(() => {
-        if (currentPathAnchor != null) {
-            setAnchorInView(currentPathAnchor);
-            anchorJustSet = true;
-            clearTimeout(anchorJustSetTimeout);
-            anchorJustSetTimeout = window.setTimeout(() => {
-                anchorJustSet = false;
-            }, 150);
-        }
-    }, [currentPathAnchor]);
-
-    useEffect(() => {
-        if (typeof window === "undefined") {
-            return;
-        }
-
-        const handleScroll = () => {
-            const scrollY = window.scrollY;
-
-            // when the user scrolls to the very top of the page, set the anchorInView to the first anchor
-            if (scrollY === 0) {
-                setAnchorInView(allAnchors[0]);
-                return;
-            }
-
-            // when the user scrolls to the very bottom of the page, set the anchorInView to the last anchor
-            if (window.innerHeight + scrollY >= document.body.scrollHeight) {
-                setAnchorInView(allAnchors[allAnchors.length - 1]);
-                return;
-            }
-
-            // when the user scrolls down, check if an anchor has just scrolled up just past 40% from the top of the viewport
-            // if so, set the anchorInView to that anchor
-
-            if (!anchorJustSet) {
-                for (let i = allAnchors.length - 1; i >= 0; i--) {
-                    const anchor = allAnchors[i];
-                    if (anchor == null) {
-                        continue;
-                    }
-                    const element = document.getElementById(anchor);
-                    if (element != null) {
-                        const { bottom } = element.getBoundingClientRect();
-                        if (bottom < window.innerHeight * 0.5) {
-                            setAnchorInView(anchor);
-                            break;
-                        }
-                    }
-                }
-            }
-        };
-
-        window.addEventListener("scroll", handleScroll);
-
-        return () => {
-            window.removeEventListener("scroll", handleScroll);
-        };
-    }, [allAnchors, tableOfContents]);
+    const allAnchorsAtom = useMemoOne(() => createTocAtom(tableOfContents), [tableOfContents]);
+    const anchorInViewAtom = useMemoOne(() => createAnchorInViewAtom(allAnchorsAtom), [allAnchorsAtom]);
+    const anchorInView = useAtomValue(anchorInViewAtom);
 
     return (
         <>
@@ -178,13 +115,3 @@ export const TableOfContents: React.FC<TableOfContents.Props> = ({ className, ta
         </>
     );
 };
-
-export interface TableOfContentsItem {
-    simpleString: string;
-    anchorString: string;
-    children: TableOfContentsItem[];
-}
-
-function getAllAnchorStrings(tableOfContents: TableOfContentsItem[]): string[] {
-    return tableOfContents.flatMap((item) => [item.anchorString, ...getAllAnchorStrings(item.children)]);
-}
