+++ packages/ui/app/src/mdx/components/layout/OverviewLayout.tsx
@@ -3,9 +3,10 @@ import { FernScrollArea } from "@fern-ui/components";
 import type { ElementContent } from "hast";
 import type { MdxJsxFlowElementHast } from "mdast-util-mdx-jsx";
 import { ReactElement, ReactNode } from "react";
+import type { TableOfContentsItem } from "../../../atoms";
 import { EditThisPageButton } from "../../../components/EditThisPage";
 import { PageHeader } from "../../../components/PageHeader";
-import { TableOfContents, TableOfContentsItem } from "../../../components/TableOfContents";
+import { TableOfContents } from "../../../components/TableOfContents";
 import { Feedback } from "../../../feedback/Feedback";
 import { BuiltWithFern } from "../../../sidebar/BuiltWithFern";
 import { toAttribute } from "../../plugins/utils";
